clean up and move examples to lib
