+++ module2/mutable_fruit_salad/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mutable_fruit_salad"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2021"
 
 [dependencies]
